Moved function html_constructor() from interface.rs to new file htmlconstructor.rs
