+++ components/script/dom/bindings/codegen/CodegenRust.py
@@ -5717,7 +5717,7 @@ def generate_imports(config, cgthings, descriptors, callbacks=None, dictionaries
         'dom::bindings::interface::define_guarded_constants',
         'dom::bindings::interface::define_guarded_methods',
         'dom::bindings::interface::define_guarded_properties',
-        'dom::bindings::interface::html_constructor',
+        'dom::bindings::htmlconstructor::html_constructor',
         'dom::bindings::interface::is_exposed_in',
         'dom::bindings::interface::pop_current_element_queue',
         'dom::bindings::interface::push_new_element_queue',
